backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic-settings>=2.4.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
cryptography>=42.0.0
pycryptodome>=3.20.0

# Request/Form handling
python-multipart>=0.0.9
email-validator>=2.2.0

# AI/LLM - Core
google-genai>=1.0.0
httpx>=0.27.0

# File Storage
boto3>=1.34.0
PyMuPDF>=1.24.0

# ML/Scientific Tools - Enhanced Features
numpy>=1.24.0
scikit-learn>=1.3.0

# Optional: Advanced ML models (comment out if not needed for basic usage)
# These provide enhanced accuracy but require more resources
sentence-transformers>=2.2.0  # SPECTER2/SciBERT for journal recommendations
transformers>=4.35.0  # RoBERTa for AI writing detection
torch>=2.0.0  # Required by transformers (CPU version sufficient)
huggingface-hub>=0.20.0  # Model downloads + authentication

# API Wrappers - Citation & Retraction Checking
pyalex>=0.13  # OpenAlex API wrapper
habanero>=1.2.0  # Crossref API wrapper
